+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [#30] use separate config map for maintenance mode
+- [#30] bump version of ces-common-lib to v0.3.0
 
 ## [v0.6.0] - 2025-10-09
 ### Added